# packages/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification for Sistema de Controle de Visitantes
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/logout:
    post:
      operationId: logout
      tags: [auth]
      summary: Logout
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Logged out
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /auth/me:
    get:
      operationId: getCurrentUser
      tags: [auth]
      summary: Get current user
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
  /auth/change-password:
    post:
      operationId: changePassword
      tags: [auth]
      summary: Change the authenticated user's password
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChangePasswordInput"
      responses:
        "200":
          description: Password changed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"

  # ── Users ─────────────────────────────────────────────────────────────────
...
  /sectors:
    get:
      operationId: listSectors
      tags: [sectors]
      summary: List sectors
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: search
          schema:
            type: string
        - in: query
          name: status
          schema:
            type: string
            enum: [active, inactive]
      responses:
        "200":
          description: List of sectors
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Sector"
    post:
      operationId: createSector
      tags: [sectors]
      summary: Create sector
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SectorInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Sector"
  /sectors/{id}:
    get:
      operationId: getSector
      tags: [sectors]
      summary: Get sector by ID
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/IdParam"
      responses:
        "200":
          description: Sector
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Sector"
    patch:
      operationId: updateSector
      tags: [sectors]
      summary: Update sector
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/IdParam"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SectorUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Sector"
    delete:
      operationId: deleteSector
      tags: [sectors]
      summary: Delete sector
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/IdParam"
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"

  # ── Visitors ──────────────────────────────────────────────────────────────
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    IdParam:
      in: path
      name: id
      required: true
      schema:
        type: integer
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required:
        - success

    # ── Auth ──────────────────────────────────────────────────────────────
    LoginInput:
      type: object
      required: [login, password]
      properties:
        login:
          type: string
        password:
          type: string
    AuthResponse:
      type: object
      required: [token, user]
      properties:
        token:
          type: string
        user:
          $ref: "#/components/schemas/User"

    # ── User ──────────────────────────────────────────────────────────────
    UserPermissions:
      type: object
      required:
        [
          viewDashboard,
          viewDashboardCharts,
          viewVisits,
          registerVisit,
          checkoutVisit,
          editVisit,
          cancelVisit,
          reprintLabel,
          viewVisitors,
          createVisitor,
          editVisitorName,
          editVisitorCpf,
          editVisitorBirthDate,
          editVisitorPhone,
          editVisitorCompany,
          editVisitorCity,
          accessServiceCenter,
          viewReports,
          manageSectors,
          manageSettings,
          viewAudit,
        ]
      properties:
        viewDashboard:
          type: boolean
        viewDashboardCharts:
          type: boolean
        viewVisits:
          type: boolean
        registerVisit:
          type: boolean
        checkoutVisit:
          type: boolean
        editVisit:
          type: boolean
        cancelVisit:
          type: boolean
        reprintLabel:
          type: boolean
        viewVisitors:
          type: boolean
        createVisitor:
          type: boolean
        editVisitorName:
          type: boolean
        editVisitorCpf:
          type: boolean
        editVisitorBirthDate:
          type: boolean
        editVisitorPhone:
          type: boolean
        editVisitorCompany:
          type: boolean
        editVisitorCity:
          type: boolean
        accessServiceCenter:
          type: boolean
        viewReports:
          type: boolean
        manageSectors:
          type: boolean
        manageSettings:
          type: boolean
        viewAudit:
          type: boolean
    User:
      type: object
      required:
        [
          id,
          name,
          login,
          role,
          status,
          mustChangePassword,
          permissions,
          createdAt,
        ]
      properties:
        id:
          type: integer
        name:
          type: string
        login:
          type: string
        role:
          type: string
          enum: [admin, receptionist, attendant]
        sectorId:
          type: ["integer", "null"]
        status:
          type: string
          enum: [active, inactive]
        mustChangePassword:
          type: boolean
        permissions:
          $ref: "#/components/schemas/UserPermissions"
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: ["string", "null"]
          format: date-time
...
    ChangePasswordInput:
      type: object
      required: [currentPassword, newPassword]
      properties:
        currentPassword:
          type: string
        newPassword:
          type: string
          minLength: 8
...
    Sector:
      type: object
      required:
        [
          id,
          name,
          abbreviation,
          secretariat,
          status,
          queueEnabled,
          usesDesks,
          createdAt,
        ]
      properties:
        id:
          type: integer
        name:
          type: string
        abbreviation:
          type: string
        secretariat:
          type: string
        status:
          type: string
          enum: [active, inactive]
        queueEnabled:
          type: boolean
        usesDesks:
          type: boolean
        createdAt:
          type: string
          format: date-time
    SectorInput:
      type: object
      required: [name, abbreviation, secretariat]
      properties:
        name:
          type: string
        abbreviation:
          type: string
        secretariat:
          type: string
        status:
          type: string
          enum: [active, inactive]
          default: active
        queueEnabled:
          type: boolean
          default: false
        usesDesks:
          type: boolean
          default: false
    SectorUpdate:
      type: object
      properties:
        name:
          type: string
        abbreviation:
          type: string
        secretariat:
          type: string
        status:
          type: string
          enum: [active, inactive]
        queueEnabled:
          type: boolean
        usesDesks:
          type: boolean

    # ── Visitor ───────────────────────────────────────────────────────────
...