# packages/api-zod/src/generated/types/userPermissions.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * API specification for Sistema de Controle de Visitantes
 * OpenAPI spec version: 0.1.0
 */

export interface UserPermissions {
  viewDashboard: boolean;
  viewDashboardCharts: boolean;
  viewVisits: boolean;
  registerVisit: boolean;
  checkoutVisit: boolean;
  editVisit: boolean;
  cancelVisit: boolean;
  reprintLabel: boolean;
  viewVisitors: boolean;
  createVisitor: boolean;
  editVisitorName: boolean;
  editVisitorCpf: boolean;
  editVisitorBirthDate: boolean;
  editVisitorPhone: boolean;
  editVisitorCompany: boolean;
  editVisitorCity: boolean;
  accessServiceCenter: boolean;
  viewReports: boolean;
  manageSectors: boolean;
  manageSettings: boolean;
  viewAudit: boolean;
}